Redesign note folding as a declarative Foldable component
Replace the fold attachment and ViewMore component, which imperatively
mounted a component into the observed element and synced styles through
a callback, classList and a global CSS class, with a single Foldable
wrapper component. The ResizeObserver now only updates an overflowing
flag, the view-more button is rendered declaratively, and all styles
are component scoped.
