--- web/src/lib/components/shared/ViewMore.svelte
@@ -1,63 +1,0 @@
-<script lang="ts">
-	import { tick } from 'svelte';
-	import { _ } from 'svelte-i18n';
-
-	interface Props {
-		onclick?: () => void | Promise<void>;
-		onfold?: (folded: boolean) => void;
-	}
-
-	let { onclick = toggle, onfold }: Props = $props();
-
-	let element = $state<HTMLDivElement>();
-	let folded = $state(true);
-
-	$effect(() => {
-		onfold?.(folded);
-	});
-
-	async function toggle(): Promise<void> {
-		folded = !folded;
-		if (folded) {
-			await tick();
-			element?.parentElement?.scrollIntoView({
-				behavior: 'smooth',
-				block: 'nearest'
-			});
-		}
-	}
-</script>
-
-<div class:folded bind:this={element}>
-	<button {onclick} class="rounded-button small">
-		{#if folded}
-			{$_('fold.unfold')}
-		{:else}
-			{$_('fold.fold')}
-		{/if}
-	</button>
-</div>
-
-<style>
-	:global(.folded-container) {
-		max-height: var(--fold-max-height, 20rem);
-		overflow: hidden;
-		position: relative;
-	}
-
-	div {
-		width: 100%;
-		padding: 0.5rem;
-		text-align: center;
-	}
-
-	div.folded {
-		position: absolute;
-		bottom: 0;
-		background: linear-gradient(transparent, var(--surface));
-	}
-
-	button {
-		background-color: var(--surface);
-	}
-</style>
